+++ Sources/ParseSwift/Objects/ParseUser+combine.swift
@@ -155,7 +155,7 @@ public extension ParseUser {
     /**
      Saves the `ParseUser` *asynchronously* and publishes when complete.
 
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -163,18 +163,18 @@ public extension ParseUser {
      - important: If an object saved has the same objectId as current, it will automatically update the current.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
      client-side checks are disabled. Developers are responsible for handling such cases.
     */
     func savePublisher(options: API.Options = [],
-                       isIgnoreCustomObjectIdConfig: Bool = false) -> Future<Self, ParseError> {
+                       ignoringCustomObjectIdConfig: Bool = false) -> Future<Self, ParseError> {
         Future { promise in
-            self.save(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+            self.save(ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig,
                       options: options,
                       completion: promise)
         }
@@ -263,7 +263,7 @@ public extension Sequence where Element: ParseUser {
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
      prevents the transaction from completing, then none of the objects are committed to the Parse Server database.
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -275,9 +275,9 @@ public extension Sequence where Element: ParseUser {
      the transactions can fail.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
@@ -285,12 +285,12 @@ public extension Sequence where Element: ParseUser {
     */
     func saveAllPublisher(batchLimit limit: Int? = nil,
                           transaction: Bool = ParseSwift.configuration.isUsingTransactions,
-                          isIgnoreCustomObjectIdConfig: Bool = false,
+                          ignoringCustomObjectIdConfig: Bool = false,
                           options: API.Options = []) -> Future<[(Result<Self.Element, ParseError>)], ParseError> {
         Future { promise in
             self.saveAll(batchLimit: limit,
                          transaction: transaction,
-                         isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+                         ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig,
                          options: options,
                          completion: promise)
         }