+++ Sources/ParseSwift/Objects/ParseUser.swift
@@ -872,13 +872,13 @@ extension ParseUser {
      - important: If an object saved has the same objectId as current, it will automatically update the current.
     */
     public func save(options: API.Options = []) throws -> Self {
-        try save(isIgnoreCustomObjectIdConfig: false, options: options)
+        try save(ignoringCustomObjectIdConfig: false, options: options)
     }
 
     /**
      Saves the `ParseUser` *synchronously* and throws an error if there's an issue.
 
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -887,9 +887,9 @@ extension ParseUser {
      - important: If an object saved has the same objectId as current, it will automatically update the current.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
@@ -897,7 +897,7 @@ extension ParseUser {
      - note: The default cache policy for this method is `.reloadIgnoringLocalCacheData`. If a developer
      desires a different policy, it should be inserted in `options`.
     */
-    public func save(isIgnoreCustomObjectIdConfig: Bool,
+    public func save(ignoringCustomObjectIdConfig: Bool,
                      options: API.Options = []) throws -> Self {
         var childObjects: [String: PointerType]?
         var childFiles: [UUID: ParseFile]?
@@ -918,7 +918,7 @@ extension ParseUser {
             throw error
         }
 
-        let result: Self = try saveCommand(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig)
+        let result: Self = try saveCommand(ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig)
             .execute(options: options,
                      childObjects: childObjects,
                      childFiles: childFiles)
@@ -929,7 +929,7 @@ extension ParseUser {
     /**
      Saves the `ParseUser` *asynchronously* and executes the given callback block.
 
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -939,9 +939,9 @@ extension ParseUser {
      - important: If an object saved has the same objectId as current, it will automatically update the current.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
@@ -950,13 +950,13 @@ extension ParseUser {
      desires a different policy, it should be inserted in `options`.
     */
     public func save(
-        isIgnoreCustomObjectIdConfig: Bool = false,
+        ignoringCustomObjectIdConfig: Bool = false,
         options: API.Options = [],
         callbackQueue: DispatchQueue = .main,
         completion: @escaping (Result<Self, ParseError>) -> Void
     ) {
         command(method: .save,
-                isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+                ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig,
                 options: options,
                 callbackQueue: callbackQueue,
                 completion: completion)
@@ -1023,7 +1023,7 @@ extension ParseUser {
 
     func command(
         method: Method,
-        isIgnoreCustomObjectIdConfig: Bool = false,
+        ignoringCustomObjectIdConfig: Bool = false,
         options: API.Options,
         callbackQueue: DispatchQueue,
         completion: @escaping (Result<Self, ParseError>) -> Void
@@ -1036,7 +1036,7 @@ extension ParseUser {
                     let command: API.Command<Self, Self>!
                     switch method {
                     case .save:
-                        command = try self.saveCommand(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig)
+                        command = try self.saveCommand(ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig)
                     case .create:
                         command = self.createCommand()
                     case .replace:
@@ -1071,8 +1071,8 @@ extension ParseUser {
         }
     }
 
-    func saveCommand(isIgnoreCustomObjectIdConfig: Bool = false) throws -> API.Command<Self, Self> {
-        if ParseSwift.configuration.isAllowingCustomObjectIds && objectId == nil && !isIgnoreCustomObjectIdConfig {
+    func saveCommand(ignoringCustomObjectIdConfig: Bool = false) throws -> API.Command<Self, Self> {
+        if ParseSwift.configuration.isAllowingCustomObjectIds && objectId == nil && !ignoringCustomObjectIdConfig {
             throw ParseError(code: .missingObjectId, message: "objectId must not be nil")
         }
         if isSaved {
@@ -1266,7 +1266,7 @@ public extension Sequence where Element: ParseUser {
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
      prevents the transaction from completing, then none of the objects are committed to the Parse Server database.
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -1279,9 +1279,9 @@ public extension Sequence where Element: ParseUser {
      the transactions can fail.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
@@ -1291,7 +1291,7 @@ public extension Sequence where Element: ParseUser {
     */
     func saveAll(batchLimit limit: Int? = nil, // swiftlint:disable:this function_body_length
                  transaction: Bool = ParseSwift.configuration.isUsingTransactions,
-                 isIgnoreCustomObjectIdConfig: Bool = false,
+                 ignoringCustomObjectIdConfig: Bool = false,
                  options: API.Options = []) throws -> [(Result<Self.Element, ParseError>)] {
         var childObjects = [String: PointerType]()
         var childFiles = [UUID: ParseFile]()
@@ -1341,7 +1341,7 @@ public extension Sequence where Element: ParseUser {
 
         var returnBatch = [(Result<Self.Element, ParseError>)]()
         let commands = try map {
-            try $0.saveCommand(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig)
+            try $0.saveCommand(ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig)
         }
         let batchLimit = limit != nil ? limit! : ParseConstants.batchLimit
         try canSendTransactions(transaction, objectCount: commands.count, batchLimit: batchLimit)
@@ -1365,7 +1365,7 @@ public extension Sequence where Element: ParseUser {
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
      prevents the transaction from completing, then none of the objects are committed to the Parse Server database.
-     - parameter isIgnoreCustomObjectIdConfig: Ignore checking for `objectId`
+     - parameter ignoringCustomObjectIdConfig: Ignore checking for `objectId`
      when `ParseConfiguration.isAllowingCustomObjectIds = true` to allow for mixed
      `objectId` environments. Defaults to false.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
@@ -1378,9 +1378,9 @@ public extension Sequence where Element: ParseUser {
      the transactions can fail.
      - warning: If you are using `ParseConfiguration.isAllowingCustomObjectIds = true`
      and plan to generate all of your `objectId`'s on the client-side then you should leave
-     `isIgnoreCustomObjectIdConfig = false`. Setting
+     `ignoringCustomObjectIdConfig = false`. Setting
      `ParseConfiguration.isAllowingCustomObjectIds = true` and
-     `isIgnoreCustomObjectIdConfig = true` means the client will generate `objectId`'s
+     `ignoringCustomObjectIdConfig = true` means the client will generate `objectId`'s
      and the server will generate an `objectId` only when the client does not provide one. This can
      increase the probability of colliiding `objectId`'s as the client and server `objectId`'s may be generated using
      different algorithms. This can also lead to overwriting of `ParseObject`'s by accident as the
@@ -1391,7 +1391,7 @@ public extension Sequence where Element: ParseUser {
     func saveAll( // swiftlint:disable:this function_body_length cyclomatic_complexity
         batchLimit limit: Int? = nil,
         transaction: Bool = ParseSwift.configuration.isUsingTransactions,
-        isIgnoreCustomObjectIdConfig: Bool = false,
+        ignoringCustomObjectIdConfig: Bool = false,
         options: API.Options = [],
         callbackQueue: DispatchQueue = .main,
         completion: @escaping (Result<[(Result<Element, ParseError>)], ParseError>) -> Void
@@ -1399,7 +1399,7 @@ public extension Sequence where Element: ParseUser {
         batchCommand(method: .save,
                      batchLimit: limit,
                      transaction: transaction,
-                     isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig,
+                     ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig,
                      options: options,
                      callbackQueue: callbackQueue,
                      completion: completion)
@@ -1507,7 +1507,7 @@ public extension Sequence where Element: ParseUser {
         method: Method,
         batchLimit limit: Int?,
         transaction: Bool,
-        isIgnoreCustomObjectIdConfig: Bool = false,
+        ignoringCustomObjectIdConfig: Bool = false,
         options: API.Options,
         callbackQueue: DispatchQueue,
         completion: @escaping (Result<[(Result<Element, ParseError>)], ParseError>) -> Void
@@ -1575,7 +1575,7 @@ public extension Sequence where Element: ParseUser {
                 switch method {
                 case .save:
                     commands = try map {
-                        try $0.saveCommand(isIgnoreCustomObjectIdConfig: isIgnoreCustomObjectIdConfig)
+                        try $0.saveCommand(ignoringCustomObjectIdConfig: ignoringCustomObjectIdConfig)
                     }
                 case .create:
                     commands = map { $0.createCommand() }